Added more comments. Also added these methods: rsid_root all_rsids
Note that all risds of varA of type Docx by either calling the method
varA.all_rsids()
or by accessing the variable itself.
varA.rsids

But it is preferable to use methods to get values out of a class rather than directly accessing a variable. As that ensures that if an update is made to the class to do additional processing of a variable before returning it via a method, using the method will ensure you are getting the correct return.