Added methods to return the unique rsidRPr, rsidP, and rsidRDefault in document.xml and count of how many times unique values of each of these was in document.xml.
Also added a method to return the # of times each rsidR is in document.xml. We don't have to look for unique rsidR in document.xml, because we get that from settings.xml. Whereas for all other rsid tags, the only place they exist is in document.xml. Thus, we need to search through it to find all of them, create a unique list of them along with a count of them.