ms_word.py class was missing some paragraph tags as some were a single tag eg: <w:p ...... />, so there was no simple closing tag </w:p>. The opening and closing was in a single tag. As a precaution, added same checks for <w:t> and <w:r>.
ms_word class now also extracts the list of paraId and textId tags in the <w:p> tags, and includes those in the RSIDs worksheet along with a count of how often each unique tag/value is observed.

Parse DOCx With Class was updated to get the new values being calculated and writing those to the Excel as well. Also grouped together all elements to be written to RSIDs worksheet into a single call of the function to write instead of resetting the rows [] list each time and writing after each sub-category of RSID being calculated.