Executable version
If you don't want to fuss with Python, you can download this executable instead. It's best to open a command window in Windows and execute it from within that window rather than double clicking on it in File Explorer. Both will work. But if you execute it by double clicking on the executable, the command window will open while it's running, but as soon as the script finishes, it closes that window. So you won't see any output from the script, including the path and file names where the output files (Excel and log file) were saved.