Two bug fixes. The function __xml_extra_bytes now checks if a filename is greater than 256 (new if statement inside the first for loop), as some DOCx files were producing false positives for excessively long file names of XML files in the DOCx resulting the script not parsing that document.
The second bug fix adds a second check when looking at the timestamp of an embedded file in the DOCx archive.
if m_time == (1980, 1, 1, 0, 0, 0) or m_time == (1980, 0, 0 , 0, 0, 0):

It was noted that in some cases, a file in the archive didn't have a date but that instead of it resulting in the tuple (1980,1,1,0,0,0) it was resulting in the tuple (1980,0,0,0,0,0). This resulted in an error parsing that document.

After the above two corrections, the script successfully parsed the files that previously failed to parse.